Refactor report generation to support streaming report generation.
This commit adds the ability to stream a report to iotile.cloud
as it is being generated.  It also refactors analytics-host to
simplify how it handles and zips report files using the new
streaming infrastructure.

In particular, there is a new FileHandler class that is passed
to a LiveReport and handles the files produced by that report.