Add support for streaming report generation (#69)
* iotile-analytics-core support for streaming report generation.

This commit adds support to iotile-analytics-core needed to
support a refactor of iotile-analytics-interactive that can
generate and stream reports to the cloud without storing the
entire report in memory at once.  This is necessary to run the
report generation process on memory constrained workers.

* Update API documentation

* Refactor report generation to support streaming report generation.

This commit adds the ability to stream a report to iotile.cloud
as it is being generated.  It also refactors analytics-host to
simplify how it handles and zips report files using the new
streaming infrastructure.

In particular, there is a new FileHandler class that is passed
to a LiveReport and handles the files produced by that report.

* Update iotile-analytics-offline for compatibility.

This is needed to handle changes in how report streaming was
implemented.

* Bump iotile-analytics-interactive to 0.5.0

* Update gitignore to ignore notebooks folder

* Update docstring and fix finish_report signature

* Update release notes for iotile-analytics-offline

* Update doc url in --help text of analytics-host

* Fix minor text edits