Add support for passing a token and report_id on command line.
analytics-host now supports two more command line flags:
--token=TOKEN
--web-push-id=REPORT_ID

If --token is specified, it will immediately set the token
for CloudSession instead of prompting for a password.

If --web-push-id is specified then when uploading a report
to iotile cloud, the files will be attached to that existing
report id, rather than creating a new report id.

Closes #70